backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.9 